Swapped the deployment order of Curve AMO and SwapX AMO
