+++ contracts/deploy/sonic/009_swapx_amo.js
@@ -4,65 +4,14 @@ const {
   withConfirmation,
 } = require("../../utils/deploy");
 const addresses = require("../../utils/addresses");
-const { oethUnits } = require("../../test/helpers");
 
 module.exports = deployOnSonic(
   {
-    deployName: "009_curve_amo",
+    deployName: "009_swapx_amo",
   },
   async ({ ethers }) => {
     const { deployerAddr, strategistAddr } = await getNamedAccounts();
     const sDeployer = await ethers.provider.getSigner(deployerAddr);
-
-    // Deploy Sonic Curve AMO Strategy proxy
-    const cOSonicProxy = await ethers.getContract("OSonicProxy");
-    const cOSonicVaultProxy = await ethers.getContract("OSonicVaultProxy");
-    const cOSonicVaultAdmin = await ethers.getContractAt(
-      "OSonicVaultAdmin",
-      cOSonicVaultProxy.address
-    );
-
-    const dSonicCurveAMOStrategyProxy = await deployWithConfirmation(
-      "SonicCurveAMOStrategyProxy",
-      []
-    );
-
-    const cSonicCurveAMOStrategyProxy = await ethers.getContract(
-      "SonicCurveAMOStrategyProxy"
-    );
-
-    // Deploy Sonic Curve AMO Strategy implementation
-    const dSonicCurveAMOStrategy = await deployWithConfirmation(
-      "SonicCurveAMOStrategy",
-      [
-        [addresses.sonic.WS_OS.pool, cOSonicVaultProxy.address],
-        cOSonicProxy.address,
-        addresses.sonic.wS,
-        addresses.sonic.WS_OS.gauge,
-        addresses.sonic.childLiquidityGaugeFactory,
-        0, // The OToken (OS for Sonic) is coin 0 of the Curve OS/wS pool
-        1, // The WETH token (wS for Sonic) is coin 1 of the Curve OS/wS pool
-      ]
-    );
-    const cSonicCurveAMOStrategy = await ethers.getContractAt(
-      "SonicCurveAMOStrategy",
-      dSonicCurveAMOStrategyProxy.address
-    );
-
-    // Initialize Sonic Curve AMO Strategy implementation
-    const initData = cSonicCurveAMOStrategy.interface.encodeFunctionData(
-      "initialize(address[],uint256)",
-      [[addresses.sonic.CRV], oethUnits("0.002")]
-    );
-    await withConfirmation(
-      // prettier-ignore
-      cSonicCurveAMOStrategyProxy
-        .connect(sDeployer)["initialize(address,address,bytes)"](
-          dSonicCurveAMOStrategy.address,
-          addresses.sonic.timelock,
-          initData
-        )
-    );
 
     // Deploy a new Vault Core implementation
     const dOSonicVaultCore = await deployWithConfirmation("OSonicVaultCore", [
@@ -101,14 +50,57 @@ module.exports = deployOnSonic(
     // Initialize the Harvester
     // prettier-ignore
     await withConfirmation(
-      cHarvesterProxy
-        .connect(sDeployer)["initialize(address,address,bytes)"](
-          dHarvester.address,
-          addresses.sonic.timelock,
-          initSonicStakingStrategy
-        )
+          cHarvesterProxy
+            .connect(sDeployer)["initialize(address,address,bytes)"](
+              dHarvester.address,
+              addresses.sonic.timelock,
+              initSonicStakingStrategy
+            )
+        );
+
+    // Deploy Sonic SwapX AMO Strategy proxy
+    const cOSonicProxy = await ethers.getContract("OSonicProxy");
+    const cOSonicVaultProxy = await ethers.getContract("OSonicVaultProxy");
+    const cOSonicVaultAdmin = await ethers.getContractAt(
+      "OSonicVaultAdmin",
+      cOSonicVaultProxy.address
+    );
+    const dSonicSwapXAMOStrategyProxy = await deployWithConfirmation(
+      "SonicSwapXAMOStrategyProxy",
+      []
+    );
+    const cSonicSwapXAMOStrategyProxy = await ethers.getContract(
+      "SonicSwapXAMOStrategyProxy"
     );
 
+    // Deploy Sonic SwapX AMO Strategy implementation
+    const dSonicSwapXAMOStrategy = await deployWithConfirmation(
+      "SonicSwapXAMOStrategy",
+      [
+        [addresses.sonic.SwapXWSOS.pool, cOSonicVaultProxy.address],
+        cOSonicProxy.address,
+        addresses.sonic.wS,
+        addresses.sonic.SwapXWSOS.gauge,
+      ]
+    );
+    const cSonicSwapXAMOStrategy = await ethers.getContractAt(
+      "SonicSwapXAMOStrategy",
+      dSonicSwapXAMOStrategyProxy.address
+    );
+    // Initialize Sonic Curve AMO Strategy implementation
+    const initData = cSonicSwapXAMOStrategy.interface.encodeFunctionData(
+      "initialize(address[])",
+      [[addresses.sonic.SWPx]]
+    );
+    await withConfirmation(
+      // prettier-ignore
+      cSonicSwapXAMOStrategyProxy
+        .connect(sDeployer)["initialize(address,address,bytes)"](
+          dSonicSwapXAMOStrategy.address,
+          addresses.sonic.timelock,
+          initData
+        )
+    );
     return {
       actions: [
         // 1. Upgrade Vault proxy to new VaultCore
@@ -123,27 +115,27 @@ module.exports = deployOnSonic(
           signature: "setAdminImpl(address)",
           args: [dOSonicVaultAdmin.address],
         },
-        // 3. Approve strategy on vault
+        // 3. Approve new strategy on the Vault
         {
           contract: cOSonicVaultAdmin,
           signature: "approveStrategy(address)",
-          args: [cSonicCurveAMOStrategyProxy.address],
+          args: [cSonicSwapXAMOStrategyProxy.address],
         },
         // 4. Add strategy to mint whitelist
         {
           contract: cOSonicVaultAdmin,
           signature: "addStrategyToMintWhitelist(address)",
-          args: [cSonicCurveAMOStrategyProxy.address],
+          args: [cSonicSwapXAMOStrategyProxy.address],
         },
-        // 5. Enable for Curve AMO after it has been deployed
+        // 5. Enable for SwapX AMO after it has been deployed
         {
           contract: cHarvester,
           signature: "setSupportedStrategy(address,bool)",
-          args: [cSonicCurveAMOStrategyProxy.address, true],
+          args: [cSonicSwapXAMOStrategyProxy.address, true],
         },
-        // 6. Set the Harvester on the Curve AMO strategy
+        // 6. Set the Harvester on the SwapX AMO strategy
         {
-          contract: cSonicCurveAMOStrategy,
+          contract: cSonicSwapXAMOStrategy,
           signature: "setHarvesterAddress(address)",
           args: [cHarvesterProxy.address],
         },
